messi

+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: dmaldona <dmaldona@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 19:11:24 by dmaldona          #+#    #+#             */
-/*   Updated: 2022/11/08 15:35:34 by dmaldona         ###   ########.fr       */
+/*   Updated: 2022/11/09 14:27:30 by dmaldona         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 
 // AUX FUNCTIONS FROM LIBFT
 size_t	ft_strlen(const char *s);
-size_t	ft_numlen(size_t n, char *base);
+int     ft_numlen(int n, char *base);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putchar_fd(char c, int fd);
 void	ft_putnbr_fd_base(int n, int fd, char *base);
